requirements.txt
flask
python-dotenv
gunicorn
requests
pyotp
schedule
email-validator
py3dns